+++ src/GSS.Authorization.OAuth/OAuthDefaults.cs
@@ -1,0 +1,24 @@
+namespace GSS.Authorization.OAuth
+{
+    public static class OAuthDefaults
+    {
+        public const string OAuthCallback = "oauth_callback";
+        public const string OAuthConsumerKey = "oauth_consumer_key";
+        public const string OAuthNonce = "oauth_nonce";
+        public const string OAuthParametersAbsent = "oauth_parameters_absent";
+        public const string OAuthPrefix = "oauth_";
+        public const string OAuthProblem = "oauth_problem";
+        public const string OAuthScheme = "OAuth";
+        public const string OAuthSignature = "oauth_signature";
+        public const string OAuthSignatureMethod = "oauth_signature_method";
+        public const string OAuthTimestamp = "oauth_timestamp";
+        public const string OAuthToken = "oauth_token";
+        public const string OAuthTokenSecret = "oauth_token_secret";
+        public const string OAuthVerifier = "oauth_verifier";
+        public const string OAuthVersion = "oauth_version";
+        public const string OutOfBand = "oob";
+        public const string ParameterAbsent = "parameter_absent";
+        public const string Realm = "realm";
+        public const string Version1 = "1.0";
+    }
+}